My attempt to fix IndexSizeErrors
The only way I could recreate this is if I try to insert an autocomplete
answer at the beginning of a string when it wraps to multiple lines. Now
instead of grabbing childNodes - 1, it will always be the first node (which
seems to be the TextNode for the editor anyway).